feat(): add Lang::Scss class
