+++ lib/hound.rb
@@ -9,6 +9,7 @@ require "hound/linters_instantiator"
 require "hound/languages/base"
 require "hound/languages/ruby"
 require "hound/languages/eslint"
+require "hound/languages/scss"
 require "hound/config/base"
 require "hound/config/ruby"
 require "hound/linters/base"
