+++ lib/hound/languages/ruby.rb
@@ -1,9 +1,9 @@
 module Hound
   module Lang
-    class Ruby < Base
+    class Scss < Base
       def initialize
-        @rules_url = "https://raw.githubusercontent.com/platanus/hound/platanus/config/style_guides/platanus/ruby.yml"
-        @file_name = ".rubocop.yml"
+        @rules_url = "https://raw.githubusercontent.com/platanus/hound-linters/platanus/config/scss.yml"
+        @file_name = ".scss-lint.yml"
       end
     end
   end
