+++ spec/languajes/scss_spec.rb
@@ -1,0 +1,5 @@
+describe Hound::Lang::Ruby do
+  subject { Hound::Lang::Ruby.new }
+
+  include_examples "get rules from url", "ruby"
+end
